Fix flatpak build
When building prusa-slicer as flatpak, Perl test suite is unable to build
because of many build-related errors (Perl tests have many dependencies
which have to be built with /app prefix, which is not always simple).

As we are unable to pass any flags to rpmbuild to switch off
the Perl tests, we need to switch them off for flatpak build directly in the spec file.

The second change is related to %{perl_vendorlib} and %{perl_vendorarch} macros,
which are not set during flatpak build (a build %without perltests).

When the %{flatpak} macro is not set, there should be no change in functionality.